Promote badge page
Closes #284
