+++ frontend/pages/badges.tsx
@@ -5,18 +5,51 @@ import cc0 from "/public/img/CC0.png"
 import { GetStaticProps } from "next"
 import { serverSideTranslations } from "next-i18next/serverSideTranslations"
 import { Trans, useTranslation } from "next-i18next"
+import { getLanguageName, languages } from "src/localize"
+import { useState } from "react"
+import Image from "next/image"
 
-const flathubBadge = "https://dl.flathub.org/assets/badges/flathub-badge-en.png"
-const flathubBadgeInverted =
-  "https://dl.flathub.org/assets/badges/flathub-badge-i-en.png"
-const badgeExampleCode = `<a href='https://flathub.org/apps/org.gimp.GIMP'>
-  <img width='240' alt='Download on Flathub' src='https://dl.flathub.org/assets/badges/flathub-badge-en.png'/>
-</a>`
-const badgeExampleCodeMoinMoin =
-  "[[https://flathub.org/apps/org.gimp.GIMP|{{https://dl.flathub.org/assets/badges/flathub-badge-en.png|Download on Flathub|width=240,align=middle}}]]"
+const BadgePreview = ({ locale, preferred }) => {
+  const { t } = useTranslation()
+
+  const lightPostfix = preferred ? "" : "&light"
+
+  return (
+    <div>
+      <h3 className="my-4 text-xl font-semibold">
+        {t(preferred ? "preferred-badge" : "alternative-badge")}
+      </h3>
+      <Image
+        width="240"
+        height="80"
+        alt="Download on Flathub"
+        src={`/api/badge?locale=${locale}${lightPostfix}`}
+      />
+      <h6 className="pt-2 text-xs font-normal">
+        <Trans i18nKey={"common:also-available-as-svg"}>
+          Also available in{" "}
+          <a
+            className="no-underline hover:underline"
+            href={`/api/badge?svg&locale=${locale}${lightPostfix}`}
+          >
+            svg format
+          </a>
+        </Trans>
+      </h6>
+    </div>
+  )
+}
 
 const Badges = () => {
   const { t } = useTranslation()
+
+  const [locale, setLocale] = useState("en")
+
+  const badgeExampleCode = `<a href='https://flathub.org/apps/org.gimp.GIMP'>
+    <img width='240' alt='Download on Flathub' src='${process.env.NEXT_PUBLIC_SITE_BASE_URI}/api/badge?locale=${locale}'/>
+  </a>`
+  const badgeExampleCodeMoinMoin = `[[https://flathub.org/apps/org.gimp.GIMP|{{${process.env.NEXT_PUBLIC_SITE_BASE_URI}/api/badge?locale=${locale}|Download on Flathub|width=240,align=middle}}]]`
+
   return (
     <>
       <NextSeo
@@ -26,122 +59,94 @@ const Badges = () => {
           url: `${process.env.NEXT_PUBLIC_SITE_BASE_URI}/badges`,
         }}
       />
-      <section className="max-w-11/12 mx-auto my-0 mt-12 w-11/12 space-y-4 2xl:w-[1400px] 2xl:max-w-[1400px]">
-        <h1 className="mb-8 text-4xl font-extrabold">{t("official-badges")}</h1>
-        <p>{t("badges-block")}</p>
+      <div className="flex max-w-full flex-col">
+        <section className={`flex flex-col px-[5%] md:px-[20%] 2xl:px-[30%]`}>
+          <h1 className="mt-8 mb-4 text-4xl font-extrabold">
+            {t("official-badges")}
+          </h1>
+          <p>{t("badges-block")}</p>
 
-        <div className="flex w-full flex-wrap justify-around">
-          <div>
-            <h3 className="my-4 text-xl font-semibold">
-              {t("preferred-badge")}
-            </h3>
-            <img
-              width="240"
-              height="80"
-              alt="Download on Flathub"
-              src={flathubBadge}
-            />
-            <h6 className="pt-2 text-xs font-normal">
-              <Trans i18nKey={"common:also-available-as-svg"}>
-                Also available in{" "}
-                <a
-                  className="no-underline hover:underline"
-                  href="https://dl.flathub.org/assets/badges/flathub-badge-en.svg"
-                >
-                  svg format
-                </a>
-              </Trans>
-            </h6>
-          </div>
-
-          <div>
-            <h3 className="my-4 text-xl font-semibold">
-              {t("alternative-badge")}
-            </h3>
-            <img
-              width="240"
-              height="80"
-              alt="Download on Flathub"
-              src={flathubBadgeInverted}
-            />
-            <h6 className="pt-2 text-xs font-normal">
-              <Trans i18nKey={"common:also-available-as-svg"}>
-                Also available in{" "}
-                <a
-                  className="no-underline hover:underline"
-                  href="https://dl.flathub.org/assets/badges/flathub-badge-i-en.svg"
-                >
-                  svg format
-                </a>
-              </Trans>
-            </h6>
-          </div>
-        </div>
-
-        <p
-          // this is a workaround for react not supporting it: https://github.com/facebook/react/issues/16563
-          {...{
-            "xmlns.dct": "http://purl.org/dc/terms/",
-            "xmlns.vcard": "http://www.w3.org/2001/vcard-rdf/3.0#",
-          }}
-        >
-          <a
-            rel="license"
-            href="http://creativecommons.org/publicdomain/zero/1.0/"
-          >
-            <FlathubImage src={cc0} alt="CC0" />
-          </a>
-          <br />
-          <Trans i18nKey={"common:badge-copyright"}>
-            To the extent possible under law,{" "}
-            <a
-              rel="dct:publisher"
-              className="no-underline hover:underline"
-              href="https://flathub.org/badges"
+          <div className="pt-8 flex flex-col w-full sm:flex-row sm:items-center gap-x-6 gap-y-4">
+            <div className="w-48 font-semibold">{t("switch-language")}</div>
+            <select
+              className="p-2 rounded-sm w-full sm:max-w-[240px]"
+              onChange={(e) => setLocale(e.target.value)}
             >
-              <span property="dct:title">Jakub Steiner</span>
-            </a>{" "}
-            has waived all copyright and related or neighboring rights to
-            <span property="dct:title">Flathub Badges</span>. This work is
-            published from: Czech Republic.
-          </Trans>
-        </p>
+              {languages
+                .filter((language) => !["ar", "fa"].includes(language))
+                .map((language) => (
+                  <option key={language} value={language}>
+                    {getLanguageName(language)}
+                  </option>
+                ))}
+            </select>
+          </div>
 
-        <h2 className="mb-6 mt-12 text-2xl font-bold">{t("code-examples")}</h2>
+          <div className="flex w-full flex-wrap justify-around pb-8 gap-6">
+            <BadgePreview locale={locale} preferred />
 
-        <div className="flex w-full flex-wrap justify-around">
+            <BadgePreview locale={locale} preferred={false} />
+          </div>
+
+          <p
+            // this is a workaround for react not supporting it: https://github.com/facebook/react/issues/16563
+            {...{
+              "xmlns.dct": "http://purl.org/dc/terms/",
+              "xmlns.vcard": "http://www.w3.org/2001/vcard-rdf/3.0#",
+            }}
+          >
+            <a
+              rel="license"
+              href="http://creativecommons.org/publicdomain/zero/1.0/"
+            >
+              <FlathubImage src={cc0} alt="CC0" />
+            </a>
+            <br />
+            <Trans i18nKey={"common:badge-copyright"}>
+              To the extent possible under law,{" "}
+              <a
+                rel="dct:publisher"
+                className="no-underline hover:underline"
+                href="https://flathub.org/badges"
+              >
+                <span property="dct:title">Jakub Steiner</span>
+              </a>{" "}
+              has waived all copyright and related or neighboring rights to
+              <span property="dct:title">Flathub Badges</span>. This work is
+              published from: Czech Republic.
+            </Trans>
+          </p>
+
+          <h2 className="mb-4 mt-12 text-2xl font-bold">
+            {t("code-examples")}
+          </h2>
+
           <div>
-            <h3 className="my-4 text-xl font-semibold">HTML</h3>
-            <CodeCopy
-              className="min-h-[180px] max-w-xs"
-              text={badgeExampleCode}
-            ></CodeCopy>
+            <h3 className="pb-4 text-xl font-semibold">HTML</h3>
+            <CodeCopy text={badgeExampleCode}></CodeCopy>
             <a href="https://flathub.org/apps/org.gimp.GIMP">
-              <FlathubImage
+              <Image
                 width={240}
                 height={80}
                 alt="Download on Flathub"
-                src={flathubBadge}
+                src={`/api/badge?locale=${locale}`}
               />
             </a>
           </div>
           <div>
-            <h3 className="my-4 text-xl font-semibold">MoinMoin Wiki</h3>
-            <CodeCopy
-              className="min-h-[180px] max-w-xs"
-              text={badgeExampleCodeMoinMoin}
-            ></CodeCopy>
+            <h3 className="pt-8 pb-4 text-xl font-semibold">MoinMoin Wiki</h3>
+            <CodeCopy text={badgeExampleCodeMoinMoin}></CodeCopy>
             <a href="https://flathub.org/apps/org.gimp.GIMP">
-              <img
+              <Image
                 width={240}
                 height={80}
                 alt="Download on Flathub"
-                src={flathubBadge}
+                src={`/api/badge?locale=${locale}`}
               />
             </a>
           </div>
-        </div>
-      </section>
+        </section>
+      </div>
     </>
   )
 }
